+++ src/Microsoft.DotNet.Build.Tasks/UpdateProjectDependencies.cs
@@ -2,7 +2,6 @@
 using Microsoft.Build.Utilities;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using NuGet.Versioning;
 using System;
 using System.IO;
 using System.Linq;
@@ -51,68 +50,6 @@ namespace Microsoft.DotNet.Build.Tasks
             }
 
             return !Log.HasLoggedErrors;
-        }
-    }
-
-    public class UpdatePrereleaseDependencies : UpdateProjectDependencies
-    {
-        [Required]
-        public string OldPrerelease { get; set; }
-
-        [Required]
-        public string NewPrerelease { get; set; }
-
-        public override bool UpdatePackage(JProperty package, string projectJsonPath)
-        {
-            var dependencyIdentifier = package.Name;
-            var dependencyVersionRange = VersionRange.Parse(package.Value.ToObject<string>());
-            var dependencyVersion = dependencyVersionRange.MinVersion;
-
-            if (dependencyVersion.Release == OldPrerelease)
-            {
-                Log.LogMessage(
-                    "Changing {0} {1} to {2} in {3}",
-                    dependencyIdentifier,
-                    dependencyVersion,
-                    NewPrerelease,
-                    projectJsonPath);
-
-                package.Value = dependencyVersionRange.OriginalString.Replace(OldPrerelease, NewPrerelease);
-                return true;
-            }
-            return false;
-        }
-    }
-
-    public class UpdatePackageDependencyVersion : UpdateProjectDependencies
-    {
-        [Required]
-        public string PackageId { get; set; }
-
-        [Required]
-        public string OldVersion { get; set; }
-
-        [Required]
-        public string NewVersion { get; set; }
-
-        public override bool UpdatePackage(JProperty package, string projectJsonPath)
-        {
-            var dependencyIdentifier = package.Name;
-            var dependencyVersion = package.Value.ToObject<string>();
-
-            if (dependencyIdentifier == PackageId && dependencyVersion == OldVersion)
-            {
-                Log.LogMessage(
-                    "Changing {0} {1} to {2} in {3}",
-                    dependencyIdentifier,
-                    dependencyVersion,
-                    NewVersion,
-                    projectJsonPath);
-
-                package.Value = NewVersion;
-                return true;
-            }
-            return false;
         }
     }
 }
